refactor(game): extract lifecycle.ts standalone transitions (Phase 4 of #132)
Move the standalone lifecycle transitions rollForStart and restoreState
into src/Game/lifecycle.ts as free functions. Public Game.* statics
remain thin delegating wrappers (external ai/api/client callers
unchanged). rollForStart self-recurses on a tie via the module-local
function (Rule A). Removed two now-dead imports from index.ts
(RESTORABLE_GAME_STATE_KINDS, BackgammonPlayerRollingForStart).

Game.initialize and Game.createNewGame stay on the class: they drive the
new Game() constructor path (PositionReconstructor, gnuPositionId getter)
and the overloaded initialize signature.

Faithful move; pre-existing casts preserved. Characterization tests for
restoreState (7) added in the prior commit stay green.

No behavior change. Core: 475 passed / 12 skipped, tsc -b clean.

Refs #132, #134
